+++ turses/api.py
@@ -26,6 +26,10 @@ except ImportError:
 DEFAULT_CACHE = object()
 
 
+class RateLimitExceededException(Exception):
+    pass
+
+
 # TODO document every method and give it arguments according to the API methods
 class TwitterApi(object):
     """
@@ -971,8 +975,11 @@ class PythonTwitterApi(BaseApi, TwitterApi):
             self._cache = cache
 
     def verify_credentials(self, *args, **kwargs):
-        # TODO
-        user = self.VerifyCredentials()
+        try:
+            user = self.VerifyCredentials()
+        except TwitterError:
+            # rate limit exceeded
+            raise RateLimitExceededException
         return User(user.screen_name)
 
     def _to_statuses(self, statuses):