+++ turses/cli.py
@@ -4,7 +4,6 @@
 #       Licensed under the GPL License. See LICENSE.txt for full details.     #
 ###############################################################################
 
-
 from .util import parse_arguments
 from .config import Configuration
 from .controller import Turses
@@ -12,6 +11,9 @@ from .ui.curses import CursesInterface
 
 
 def main():
-    configuration = Configuration(parse_arguments())
-    ui = CursesInterface()
-    Turses(configuration, ui)
+    try:
+        configuration = Configuration(parse_arguments())
+        ui = CursesInterface()
+        Turses(configuration, ui)
+    except KeyboardInterrupt:
+        exit(0)